added auth to blog/delete validate user can only delete their own blogs

+++ bloglist/controllers/blogs.js
@@ -61,9 +61,23 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   const { id } = request.params;
+  const { token } = request;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!decodedToken.id) return response.status(401).json({ error: 'token missing or invalid' })
 
-  await Blog.findByIdAndRemove(id);
-  response.status(204).end();
+  let user = await User.findById(decodedToken.id);
+  let blog = await Blog.findById(id)
+
+  if (`${user._id}` === `${blog.user}`)  {
+    const noteToRemove = await Blog.findByIdAndRemove(id)
+    if (noteToRemove) {
+      return response.status(204).end()
+    }
+  } else {
+    return response.status(401).json({
+      error: 'operation not allowed'
+    }).end();
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
